unirec: BUGFIX skip translation for missing fields
When an UniRec field is not in the UniRec template, it is not
possible to write translated value into message.  Therefore,
it is useless to translate fields that are not in template at all.

This patch removes the translate function from the map for fields that
are not present in the template.  Check was added too to skip processing
the IPFIX field.