unirec: BUGFIX check "todo" list of fields
When the list of IPFIX fields of a record is processed, the plugin
checks the list of remaining UniRec fields that are required.
If some required UniRec field is not filled, the message cannot be
sent.

This patch fixes the condition that distinguishes end of the list with
everything filled and not-filled-yet field.