Core: process remaining flow records on termination (SIGINT, SIGTERM)
Until now, when an external signal was received, the collector was quickly
terminated and the received but not yet processed flow records were discarded.

Now, when a signal is received, there is a gradual termination where the
reception of new flow records is terminated but all records received so far
are fully processed. Only if the termination signal arrives a second time
the processing of unprocessed records will be also interrupted.