core: fix includes in utils.c
#include <ipfixcol2.h> resulted in including a file in which
be64toh/htobe64 functions were used. This would clash with the defines
present here that are necessary to get the wanted strerror_r function,
resulting in the be64toh/htobe64 functions being undefined.

This was fixed by not including ipfixcol2.h as a whole, but only the
necessary part, as well as adding explicit includes for the system
headers that were brought by ipfixcol2.h.