FDS Output: allow more descriptive error messages when std::exception is thrown
Previously, we would only print exception description if the exception
was of type FDS_exception, otherwise we would fail with "Unknown error",
which does not help with figuring out what the problem was at all.

We ourselves throw descriptive exceptions of type std::runtime_error for
example in case of Config parsing errors, that would never get printed
in the original state, and instead we'd still fail with "Unknown error".