TCP input: add missing CMake dependency on Lz4
This seemed to mostly work because libfds depends on Lz4 and so we would mostly
find Lz4.h in the default system include path on Linux.
On FreeBSD though, Lz4.h is in a path that is not a part of the includes that
CMake would use by default, which revealed this issue.