Forwarder: clear buffers on connection lost
Unfinished transfers would remain in buffers when we lost connection.

In case of TCP, this is a problem. A message might be half-way send when
the connection drops, resulting in the other half being sent after the
TCP connection reestabilishes. The collector would interpret this as an
entirely new message, resulting in bogus data being decoded.